修改DWORD为unsigned long类型

+++ ImageBasedIibrary/ImageBasedIibrary/tch_track.h
@@ -42,13 +42,13 @@ extern "C" {
 //¼ÆÊ±Æ÷
 typedef struct TrackTimer
 {
-	DWORD start;
-	DWORD finish;
+	unsigned long start;
+	unsigned long finish;
 	//clock_t start;
 	//clock_t finish;
 	/*double timeLast;
 	double timeNow;*/
-	double deltaTime;
+	unsigned long deltaTime;
 }Tch_Timer_t;
 
 //Ô¤ÖÃÎ»¿é
